@@ -13,10 +13,11 @@
     "build": "npm run check_versions && tsc",
     "watch": "tsc -w",
     "tslint": "tslint -p tsconfig.json",
-    "test": "jest --forceExit --detectOpenHandles",
+    "test": "jest --forceExit --detectOpenHandles /tests/rest/integration.test.ts",
     "test:watch": "npm run test -- --watchAll",
     "validate": "npm i && npm run tslint && npm run test && npm run clean-build",
-    "check_versions": "node ../../.scripts/check_package_version.js"
+    "check_versions": "node ../../.scripts/check_package_version.js",
+    "migrate-jest": "ts-jest config:migrate jest.config.js"
   },
   "keywords": [
     "signalwire",
@@ -36,11 +37,10 @@
   "homepage": "https://github.com/signalwire/signalwire-node",
   "license": "MIT",
   "dependencies": {
-    "@types/twilio": "0.0.10",
+    "@signalwire/compatibility-api": "^3.0.5",
     "dotenv": "^8.2.0",
     "faye-websocket": "^0.11.3",
     "loglevel": "^1.6.8",
-    "twilio": "~3.73",
     "uuid": "^7.0.3"
   },
   "overrides": {
@@ -49,16 +49,16 @@
   },
   "devDependencies": {
     "@types/dotenv": "^8.2.0",
-    "@types/jest": "^25.2.1",
-    "@types/node": "^13.11.1",
-    "@types/uuid": "^7.0.2",
-    "graphql": "^0.12.0",
-    "jest": "^25.3.0",
-    "mockttp": "^0.19.3",
-    "nock": "^12.0.3",
-    "ts-jest": "^25.3.1",
-    "ts-loader": "^6.2.2",
-    "tslint": "^6.1.1",
-    "typescript": "^3.8.3"
+    "@types/jest": "^29.5.8",
+    "@types/node": "^20.9.1",
+    "@types/uuid": "^9.0.7",
+    "graphql": "^15.8.0",
+    "jest": "^29.7.0",
+    "mockttp": "^3.9.4",
+    "nock": "^13.3.8",
+    "ts-jest": "^29.1.1",
+    "ts-loader": "^9.5.1",
+    "tslint": "^6.1.3",
+    "typescript": "^5.2.2"
   }
 }
