@@ -7,7 +7,6 @@ import Event from '../Event'
 export class Tap extends Controllable {
   public eventType: string = CallNotification.Tap
   public method: string = CallMethod.Tap
-  public controlId: string = this.controlId
 
   constructor(public call: Call, public tap: IRelayCallingTapTap, public device: IRelayCallingTapDevice) {
     super(call)
@@ -23,7 +22,7 @@ export class Tap extends Controllable {
     }
   }
 
-  get sourceDevice(): IRelayCallingTapDevice {
+  get sourceDevice(): IRelayCallingTapDevice | null {
     if (!this._executeResult) {
       return null
     }
